### 2023-06-21 #5579 @editorialbot
[PRE REVIEW]: EdSurvey: an R package to analyze Large-scale educational assessments data
**Submitting author:** @pdbailey0 (<a href="http://orcid.org/0000-0003-0989-8729">Paul Bailey</a>)
**Repository:** https://github.com/American-Institutes-for-Research/EdSurvey/
**Branch with paper.md** (empty if default branch): joss
**Version:** 4.0.1
