### 2023-06-29 409387a
update tests to dropOmittedLevels
