+++ R/readPIRLS.R
@@ -416,6 +416,7 @@ convertPIRLSYearCode <- function(yrCode) {
   yrTest[yrTest %in% "r2"] <- 2006
   yrTest[yrTest %in% "r3"] <- 2011
   yrTest[yrTest %in% c("r4", "l2")] <- 2016
+  yrTest[yrTest %in% "r5"] <- 2021
 
   return(yrTest)
 }
@@ -423,8 +424,8 @@ convertPIRLSYearCode <- function(yrCode) {
 getPIRLSYearCodes <- function() {
   # retrieve the TIMMS years based on their filenaming structure
 
-  yrVals <- c("r1", "r2", "r3", "r4", "l2")
-  names(yrVals) <- c(2001, 2006, 2011, 2016, 2016)
+  yrVals <- c("r1", "r2", "r3", "r4", "l2", "r5")
+  names(yrVals) <- c(2001, 2006, 2011, 2016, 2016, 2021)
 
   return(yrVals)
 }
@@ -548,13 +549,13 @@ processPIRLS <- function(dataFolderPath, countryCode, fnames, fileYrs, forceRere
     if (nrow(stuDF1) != nrow(mm)) {
       stop(paste0(
         "Failed consistency check for filetype ", sQuote("asa"), " country code ", sQuote(tolower(countryCode)), ". ",
-        "Please email EdSurvey.help@air.org for assistance."
+        "Please report this to https://github.com/American-Institutes-for-Research/EdSurvey/issues for assistance."
       ))
     }
     if (nrow(stuDF2) != nrow(mm)) {
       stop(paste0(
         "Failed consistency check for filetype ", sQuote("asg"), " country code ", sQuote(tolower(countryCode)), ". ",
-        "Please email EdSurvey.help@air.org for assistance."
+        "Please report this to https://github.com/American-Institutes-for-Research/EdSurvey/issues for assistance."
       ))
     }
     if (min(is.na(ash)) == 0) {
@@ -577,13 +578,13 @@ processPIRLS <- function(dataFolderPath, countryCode, fnames, fileYrs, forceRere
       if (nrow(stuDF1) != nrow(mm)) {
         stop(paste0(
           "Failed consistency check for filetype ", sQuote("ash"), " country code ", sQuote(tolower(countryCode)), ". ",
-          "Please email EdSurvey.help@air.org for assistance."
+          "Please report this to https://github.com/American-Institutes-for-Research/EdSurvey/issues for assistance."
         ))
       }
       if (nr != nrow(mm)) {
         stop(paste0(
           "Failed consistency check for filetype ", sQuote("ash"), " country code ", sQuote(tolower(countryCode)), ". ",
-          "Please email EdSurvey.help@air.org for assistance."
+          "Please report this to https://github.com/American-Institutes-for-Research/EdSurvey/issues for assistance."
         ))
       }
     } else {
@@ -615,7 +616,7 @@ processPIRLS <- function(dataFolderPath, countryCode, fnames, fileYrs, forceRere
       if (nr != nrow(mm)) {
         stop(paste0(
           "Failed consistency check for filetype ", sQuote("asr"), " country code ", sQuote(tolower(countryCode)), ". ",
-          "Please email EdSurvey.help@air.org for assistance."
+          "Please report this to https://github.com/American-Institutes-for-Research/EdSurvey/issues for assistance."
         ))
       }
       mm <- mm[ , names(mm)[!grepl("\\.junk$", names(mm))]]
@@ -654,7 +655,7 @@ processPIRLS <- function(dataFolderPath, countryCode, fnames, fileYrs, forceRere
     if (nrow(stuTeachDF) != nrow(mm)) {
       stop(paste0(
[…]
