+++ tests/testthat/test-40-ELS.R
@@ -7,13 +7,6 @@ options(useFancyQuotes = FALSE)
 
 
 source("REF-7-ELS.R") # has REF output in it
-if (!exists("edsurveyHome")) {
-  if (Sys.info()[["sysname"]] == "Windows") {
-    edsurveyHome <- "C:/EdSurveyData/"
-  } else {
-    edsurveyHome <- "~/EdSurveyData/"
-  }
-}
 
 if (!dir.exists(edsurveyHome)) {
   dir.create(edsurveyHome)
