+++ tests/testthat/test-50-pisa.R
@@ -6,13 +6,6 @@ options(width = 500)
 options(useFancyQuotes = FALSE)
 source("REF-4-pisa.R") # has REF output in it
 
-if (!exists("edsurveyHome")) {
-  # if (Sys.info()[['sysname']] == "Windows") {
-  #   edsurveyHome <- "C:/EdSurveyData/"
-  # } else {
-  edsurveyHome <- "~/EdSurveyData/"
-  # }
-}
 
 if (!dir.exists(edsurveyHome)) {
   dir.create(edsurveyHome)

[…]
