### 2023-08-18 #5579 @oliviaguest
Would any of you @lebebr01 @wjakethompson @kalendar be happy to review this paper for JOSS?
