### 2023-09-07 #5579 @jrosen48
Could I have six weeks! I am on parental leave for another four weeks. Thank you!

--
Joshua M. Rosenberg, Ph.D.
Associate Professor, STEM Education
Coordinator, Educational Data Science Graduate Certificate Program
The University of Tennessee, Knoxville
https://joshuamrosenberg<https://joshuamrosenberg.com/>

PI, CAREER: Creatively Reimagining Engagements with Data in Biology Learning Environments (NSF Grant No. 2239152)

Associate Editor, Journal of Statistics and Data Science Education
