### 2023-09-11 American-Institutes-for-Research/EdSurvey#62 @wjakethompson
Automate and increase test coverage
JOSS review criterion: Automated tests (openjournals/joss-reviews#5835)

Using `covr::package_coverage()`, I see only 14.88% test coverage. I think this should be substantially higher in order to adequately verify that the package functions as expected and to ensure that code is immune to bugs that may be introduced by future changes.

Additionally, it does not appear that tests are automatically run (e.g., with GitHub actions or similar). The tests would be run on CRAN when an updated version is submitted, but tests should also be run when updates are made on GitHub. Running `usethis::use_github_action("test-coverage")` will create a GitHub actions workflow that should run the tests automatically on changes to `main` and PRs.

```
covr::package_coverage()

EdSurvey Coverage: 14.88%
R/all.R: 0.00%
R/cacheMetaReqUpdate.R: 0.00%
R/contourPlot.R: 0.00%
R/copyDataToTemp.R: 0.00%
R/download_ePIRLS.R: 0.00%
R/downloadECLS_K.R: 0.00%
R/downloadELS.R: 0.00%
R/downloadHSLS.R: 0.00%
R/downloadICCS.R: 0.00%
R/downloadICILS.R: 0.00%
R/downloadNHES.R: 0.00%
R/downloadPIAAC.R: 0.00%
R/downloadPIRLS.R: 0.00%
R/downloadPISA_YAFS.R: 0.00%
R/downloadPISA.R: 0.00%
R/downloadSSOCS.R: 0.00%
R/downloadTALIS.R: 0.00%
R/downloadTIMSS.R: 0.00%
R/downloadTIMSSAdv.R: 0.00%
R/drawPVs.sdf.R: 0.00%
R/edsurveyTable2pdf.R: 0.00%
R/glm.sdf.R: 0.00%
R/head.edsurvey.data.frame.R: 0.00%
R/mixed.sdf.R: 0.00%
R/mml.scoreTIMSS.R: 0.00%
R/mml.sdf.helpers.R: 0.00%
R/mvrlm.sdf.R: 0.00%
R/nhesTable.R: 0.00%
R/parseScript_SPSS.R: 0.00%
R/percentile.R: 0.00%
R/predictGLM.R: 0.00%
R/print.edsurvey.data.frame.R: 0.00%
R/read_ePIRLS.R: 0.00%
R/readBTLS.R: 0.00%
R/readECLS_B.R: 0.00%
R/readECLS_K1998.R: 0.00%
R/readECLS_K2011.R: 0.00%
R/readELS.R: 0.00%
R/readHSB_SO.R: 0.00%
R/readHSB_SR.R: 0.00%
R/readHSLS.R: 0.00%
R/readHSTS.R: 0.00%
R/readICCS.R: 0.00%
R/readICILS.R: 0.00%
R/readNAEP_XML.R: 0.00%
R/readNHES.R: 0.00%
R/readPIAAC.R: 0.00%
R/readPIRLS.R: 0.00%
R/readPISA_YAFS.R: 0.00%
R/readPISA.R: 0.00%
R/readSSOCS.R: 0.00%
R/readTALIS.R: 0.00%
R/readTIMSS.R: 0.00%
R/readTIMSSAdv.R: 0.00%
R/readUTILS_PRINT.R: 0.00%
R/rename.sdf.R: 0.00%
R/rq.sdf.R: 0.00%
R/sameSurvey.R: 0.00%
R/showCutPoints.R: 0.00%
R/suggestWeights.R: 0.00%
R/testthat-helpers.R: 0.00%
R/readUTILS.R: 0.82%
R/checkDataClass.R: 3.57%
R/mml.paramtabs.R: 6.10%
R/mml.sdf.R: 6.92%
R/edsurvey.data.frame.list.R: 9.92%
R/waldTest.R: 10.68%
R/rbind.light.edsurvey.data.frame.R: 15.00%
R/showPlausibleValues.R: 21.52%
R/recode.sdf.R: 21.65%
R/gap.R: 23.05%
R/showWeights.R: 31.43%
R/utilities.R: 37.61%
R/edsurveyTable.R: 38.20%
R/searchSDF.R: 38.53%
R/readNAEP.r: 48.69%
R/subset.light.edsurvey.data.frame.R: 50.00%
R/getItems.R: 51.35%
R/rebindAttributes.R: 52.38%
R/lm.sdf.R: 53.36%
R/edsurvey.data.frame.R: 53.38%
R/subset.edsurvey.data.frame.list.R: 55.26%
R/summary2.R: 57.59%
R/cor.sdf.R: 58.92%
R/achievementLevels.R: 61.67%
R/getAttributes.R: 61.90%
R/descriptionOfFile.R: 64.10%
R/mml.scoreNAEP.R: 65.62%
R/merge.sdf.R: 68.18%
R/levelsSDF.R: 69.63%
R/itterateESDFL.R: 70.00%
R/varEstToCov.R: 72.34%
R/getData.R: 75.29%
R/DoFCorrection.R: 77.14%
R/showCodebook.R: 78.79%
R/mergev.R: 81.82%
R/dim.edsurvey.data.frame.R: 82.14%
R/setAttributes.R: 86.67%
R/cbind.light.edsurvey.data.frame.R: 95.65%
R/as.data.frame.light.edsurvey.data.frame.R: 100.00%
R/zzz.R: 100.00%
```
