### 2023-09-11 American-Institutes-for-Research/EdSurvey#63 @wjakethompson
Add community guidelines
JOSS review criterion: Community guidelines (openjournals/joss-reviews#5835)

The README mentions that external contributions will be considered, but no clear guidelines are provided. `usethis::use_tidy_contributing()` provides a good starting point for creating this document (modifications will be needed, but most of the content is relevant).

I would also recommended adding a code of conduct, which can be added using `usethis::use_code_of_conduct()`, and then modified as needed.
