### 2023-09-11 American-Institutes-for-Research/EdSurvey#64 @wjakethompson
Code functionality issues
JOSS review criterion: Functionality (openjournals/joss-reviews#5835)

`goodpractice` picked up several issues that may impact package functionality.

1. Test coverage (already addressed in #62)
2. Cyclomatic complexity is very high for many functions. High complexity means that code is more difficult to maintain, is less legible, and is more difficult to evaluate for the correctness of the implementation. Maybe try refactoring or modularizing function to get complexity down to more reasonable values (e.g., 20-50)?
3. `sapply()` is not type safe and may provide unexpected results in edge cases.
4. Using `seq_along()` and `seq_length()` also avoid errors in edge cases that are introduced by, for example, `1:nrow(...)`.

There are also a few other notes from `goodpractice` included below that would be helpful, but are less critical.

```
goodpractice::gp()

── GP EdSurvey ─────────────────────────────────────────────────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 14% of code lines
    are covered by test cases.

    R/achievementLevels.R:135:NA
    R/achievementLevels.R:136:NA
    R/achievementLevels.R:138:NA
    R/achievementLevels.R:156:NA
    R/achievementLevels.R:171:NA
    ... and 20346 more lines

  ✖ write short and simple functions. These functions have high cyclomatic complexity
    (>50): getData (179), gap (132), readCivEDICCS (109), calcEdsurveyTable (104), gapHelper (104),
    readTIMSS (95), mergev (92), calc.lm.sdf (87), rename.sdf (87), readPISA (86), readPIRLS (76),
    readDict (74), calc.glm.sdf (73), percentile (73), calAL (72), cor.sdf (68), readNAEP (67),
    mixed.sdf (65), calc.mvrlm.sdf (57), processTIMSS4AndNumeracy (56), fixCivED_DataLabels (54),
    readTALIS (53), validate_NAEPFilePaths (52). You can make them easier to reason about by
    encapsulating distinct steps of your function into subfunctions.
  ✖ not use "Depends" in DESCRIPTION, as it can cause name clashes, and poor interaction
    with other packages. Use "Imports" instead.
  ✖ omit "Date" in DESCRIPTION. It is not required and it gets invalid quite often. A build
    date will be added to the package when you perform `R CMD build` on it.
  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor
    windows that are about 80 characters wide. Try make your lines shorter than 80 characters

    inst/doc/introduction.R:1:81
    inst/doc/introduction.R:4:81
    inst/doc/introduction.R:7:81
    R/achievementLevels.R:3:81
    R/achievementLevels.R:5:81
    ... and 6149 more lines

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list, depending on
    the input data. Consider using vapply() instead.

    R/achievementLevels.R:227:16
    R/achievementLevels.R:238:17
    R/achievementLevels.R:244:24
    R/achievementLevels.R:246:31
    R/achievementLevels.R:247:29
    ... and 199 more lines

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...) expressions.
    They are error prone and result 1:0 if the expression on the right hand side is zero. Use
    seq_len() or seq_along() instead.

    R/achievementLevels.R:384:13
    R/achievementLevels.R:512:37
    R/achievementLevels.R:537:37
    R/achievementLevels.R:546:36
    R/achievementLevels.R:586:36
    ... and 406 more lines

  ✖ not import packages as a whole, as this can cause name clashes between the imported
    packages. Instead, import only the specific functions you need.
  ✖ avoid 'T' and 'F', as they are just variables which are set to the logicals 'TRUE' and
    'FALSE' by default, but are not reserved words and hence can be overwritten by the user.
    Hence, one should always use 'TRUE' and 'FALSE' for the logicals.

    R/lm.sdf.R:NA:NA
    R/lm.sdf.R:NA:NA
    R/lm.sdf.R:NA:NA
    R/mml.scoreNAEP.R:NA:NA

────────────────────────────────────────────────────────────────────────────────────────────────────────────────
```

```
cyclocomp::cyclocomp_package("EdSurvey")
                                       name cyclocomp
190                                 getData       179
177                                     gap       132
410                           readCivEDICCS       109
71                        calcEdsurveyTable       104
178                               gapHelper       104
432                               readTIMSS        95
285                                  mergev        92
67                              calc.lm.sdf        87
437                              rename.sdf        87
428                                readPISA        86
427                               readPIRLS        76
411                                readDict        74
66                             calc.glm.sdf        73
333                              percentile        73
65                                    calAL        72
111                                 cor.sdf        68
424                                readNAEP        67
286                               mixed.sdf        65
68                           calc.mvrlm.sdf        57
398                processTIMSS4AndNumeracy        56
170                     fixCivED_DataLabels        54
431                               readTALIS        53
486                  validate_NAEPFilePaths        52
```
