### 2023-09-11 American-Institutes-for-Research/EdSurvey#65 @wjakethompson
Example usage questions
JOSS review criterion: Example usage (openjournals/joss-reviews#5835)

Example usage in the vignette currently points to external documents. These documents do a good job of illustrating example uses, but some were create many years ago (e.g., 2018). Is the functionality in those older documents still accurate, or have updates to the package in the last 5 years made those example obsolete? On the other hand, will future changes to EdSurvey potentially make current examples irrelevant.

Ideally, each of the example uses would be included as a vignette in the package (or an article that is only rendered on the pkgdown site if they are too big for CRAN). This would ensure that all the examples still work with the current version of the package, and would also ensure that future updates don't break the example (or that the examples are updated to reflect the updated package).
