### 2023-09-11 #5835 @wjakethompson
I have completed my review, with a few outstanding issues:

1. Currently there is no LICENSE file. I have opened a PR to add this document. crossref: American-Institutes-for-Research/EdSurvey#61
2. There are currently no contributor guidelines, just a mention in the README that contributions will be considered. Specific guidelines should be added, and I would also recommend adding a code of conduct. crossref: American-Institutes-for-Research/EdSurvey#63
3. Test coverage is very low (~15%) and does not appear to be automated. I think this should be substantially higher to meet the documentation criterion. crossref: American-Institutes-for-Research/EdSurvey#62
4. A few recommendations from `goodpractice::gp()` to ensure functionality of the package works as intended. crossref: American-Institutes-for-Research/EdSurvey#64
5. Example uses are pre-rendered and often several years old, making it unclear whether these examples are still relevant, or will remain relevant. crossref: American-Institutes-for-Research/EdSurvey#65
