### 2023-09-11 American-Institutes-for-Research/EdSurvey#65 @pdbailey0
Thanks for pointing this out. EdSurvey was taking too long to build on the CRAN servers and was getting close to oversize, so we removed the vignettes to comply with a request from the CRAN maintainers to make it build at least two times faster.

The pkgdown site is an interesting alternative. We're actually moving our documentation to [a book in package down](https://naep-research.airprojects.org/R-D-Hub/the-edsurvey-e-book-is-now-available). @wjakethompson If we updated the vignette to link there would that alleviate your concerns? Or are you saying it would be part of [this site] (https://american-institutes-for-research.github.io/EdSurvey/)?
