### 2023-09-11 American-Institutes-for-Research/EdSurvey#64 @pdbailey0
Thanks for sharing these results. I'm just starting to look at this, but I'm a bit confused by this one.
```
not use "Depends" in DESCRIPTION, as it can cause name clashes, and
    poor interaction with other packages. Use "Imports" instead.
```
depends is different from imports because it attaches the package in the user space. For example, without `lfactors` in depends, the user couldn't use the `lfactors` functionality built into `EdSurvey`. I thought about this and chose the appropriate level for each package, and for most it was imports. But doesn't just dominate depends, they both have a role.

My question is, are the `goodpractice` results negotiable for you? And how does that happen. It seems a bit blunt in the above issue without a comment about this.
