### 2023-09-19 American-Institutes-for-Research/EdSurvey#65 @wjakethompson
I think in an ideal world it would be part of pkgdown site directly. If the example uses are part of the package itself, then they will be rebuilt whenever the package is updated, and throw and error if any of the changes to the package have broken the example use. This way, you would always know that the example uses were up to date with the current version of the package.

However, if EdSurvey is stable at this point and you aren't anticipating many updates in the future, then I think the bookdown site is sufficient.
