### 2023-09-25 9dd025c
update readme to reference code of conduct and contributing guideline
