+++ .github/CONTRIBUTING.md
@@ -9,8 +9,7 @@ You can fix typos, spelling mistakes, or grammatical errors in the documentation
 ## Bigger changes
 
 If you want to make a bigger change, it's a good idea to first file an issue and make sure someone from the team agrees that it’s needed.
-If you’ve found a bug, please file an issue that illustrates the bug with a minimal 
-[reprex](https://www.tidyverse.org/help/#reprex) (this will also help you write a unit test, if needed).
+If you’ve found a bug, please file an issue that illustrates the bug with a minimal reproducible example.
 
 ### Pull request process
 
@@ -22,11 +21,9 @@ If you’ve found a bug, please file an issue that illustrates the bug with a mi
 *   Make your changes, commit to git, and then create a PR by running `usethis::pr_push()`, and following the prompts in your browser.
     The title of your PR should briefly describe the change.
     The body of your PR should contain `Fixes #issue-number`.
-*  For user-facing changes, add a bullet to the top of `NEWS.md` (i.e. just below the first header). Follow the style described in <https://style.tidyverse.org/news.html>.
 
-### Code style
+### Documentation and testing
 
-*  You can use the [styler](https://CRAN.R-project.org/package=styler) package to style your code, but please don't restyle code that has nothing to do with your PR.  
 *  We use [roxygen2](https://cran.r-project.org/package=roxygen2), with [Markdown syntax](https://cran.r-project.org/web/packages/roxygen2/vignettes/rd-formatting.html), for documentation.  
 *  We use [testthat](https://cran.r-project.org/package=testthat) for unit tests. 
    Contributions with test cases included are easier to accept.  
