### 2023-09-27 d8cd3e0
update community guidelines based on comments from General Counsel
