### 2023-09-29 American-Institutes-for-Research/EdSurvey#63 @yuqiliao
Thanks for your suggestion! We've added the community guidelines as suggested, which can be seen from the [contribution section of the readme page](https://github.com/American-Institutes-for-Research/EdSurvey#contributions), hence closing this issue. Please feel free to re-open as needed.
