### 2023-10-11 #5835 @oliviaguest
> @oliviaguest  I think you intended to at tag me, so I'll respond. Yes, it is clear what I should do, I'm working on my response. EdSurvey is a pretty large and mature package. As an example, I have over a hundred `sapply`s to look at!

Indeed, that autocomplete with a migraine is not ideal. 😂
