### 2024-01-10 American-Institutes-for-Research/EdSurvey#64 @pdbailey0
@wjakethompson I spent some time working on `cyclocomp` and I'm reminded of [Campbell's law](https://en.wikipedia.org/wiki/Campbell's_law).

I moved pretty easy to understand input validation into validation helpers. This does the same as just adding a comment "start validation" and another "end validation" and so does little, if anything, for code complexity but did decrease the number cyclocomp returned. I also find packages with low cyclocomp (like `ggplot2`) nearly unreadable because of how scope is handled in R and it gets very difficult to track where variables are coming from. I get that ideally I would have paid more attention to this while developing the softare, but now I'm less sure. Is this a deal breaker for you?
