+++ R/mergev.R
@@ -62,21 +62,209 @@ mergev <- function(x, y,
     on.exit(options(o0))
   }
 
+  order <- match.arg(order)
+  params <- list(...)
+  checkMergevVars(x=x, y=y,
+                  by=by, by.x=by.x, by.y=by.y,
+                  order=order,
+                  fast=fast,
+                  merge.type.colname = merge.type.colname,
+                  showWarnings=showWarnings,
+                  params=params)
+
   # get the names x and y were called with
   x.arg.name <- paste(deparse(substitute(x)), collapse = "")
   y.arg.name <- paste(deparse(substitute(y)), collapse = "")
 
-  genVarNames <- function(x, df) {
-    while (TRUE) {
-      prev <- substr(x, 2, nchar(x))
-      y <- paste0(x, prev, sample(9, 1))
-      if (!y %in% names(df)) {
-        break
-      }
-    }
-    y
+  
+  lst <- list()
+
+  bys <- resolveMergevBy(x, y, by, by.x, by.y)
+  x <- bys$x
+  y <- bys$y
+  by.x <- bys$by.x
+  by.y <- bys$by.y
+
+  if (is.null(all.x)) {
+    all.x <- all
   }
 
+  if (is.null(all.y)) {
+    all.y <- all
+  }
+
+  checkBy(x, by.x, x.arg.name)
+  checkBy(y, by.y, y.arg.name)
+
+  for (i in 1:length(by.x)) {
+    nx <- sum(is.na(x[ , by.x[i]]))
+    ny <- sum(is.na(y[ , by.y[i]]))
+    if (showWarnings && (nx >= 1 || ny >= 1)) {
+      warning("For by variable, '", by.x[i], "' There are ", nx, " NA values on ", x.arg.name, " and ", ny, " NA values on ", y.arg.name, ".")
+    }
+  }
+
+  if (order %in% c("x", "y")) {
+    order.var <- "order"
+    while (order.var %in% c(names(x), names(y))) {
+      order.var <- paste0(order.var, sample(LETTERS, 1))
+    }
+    if (order == "x") {
+      x[ , order.var] <- 1:nrow(x)
+    } else {
+      y[ , order.var] <- 1:nrow(y)
+    }
+  } else {
+    if (order == "sort") {
+      sort <- TRUE
+    } else {
+      sort <- FALSE
+    }
+  }
+
+  tempx <- genVarNames("x", x)
+  tempy <- genVarNames("y", y)
+
+  x[ , tempx] <- c(1:nrow(x))
+  y[ , tempy] <- c(1:nrow(y))
+
+  v1 <- names(x)
+  v2 <- names(y)
+
+  # create the verbose output matrix
+  lst$merge.summary <- buildMatTop(x=x, y=y, by.x = by.x, by.y= by.y, x.arg.name=x.arg.name, y.arg.name=y.arg.name)
+
+  merge_type <- ""
+  not.one.to.one <- FALSE
+  if (lst$merge.summary[2, 1] < lst$merge.summary[1, 1]) {
+    lst$merge.type <- c("many:")
+    merge_type <- "many:"
+    not.one.to.one <- TRUE
+  } else {
+    lst$merge.type <- c("one:")
+    merge_type <- "one:"
+  }
+
+  if (lst$merge.summary[2, 2] < lst$merge.summary[1, 2]) {
+    lst$merge.type <- c(lst$merge.type, "many")
+    merge_type <- paste0(merge_type, "many")
+    not.one.to.one <- TRUE
+  } else {
+    lst$merge.type <- c(lst$merge.type, "one")
+    merge_type <- paste0(merge_type, "one")
+  }
+
+  # make output
[…]
