### 2024-01-23 American-Institutes-for-Research/EdSurvey#64 @pdbailey0
@wjakethompson, I just made a [commit to main](https://github.com/American-Institutes-for-Research/EdSurvey/commit/e9781b7e4b7dce2d2a1c4003fd01672b3912b371) that updated `mergev`, which is quite compartmentalized and well tested. It decreased the complexity to 52, but I don't see that the function is any better, more maintainable, or understandable.

I also hesitate to adjust a function like `gap` or `getData` because they are mature received extensive acceptance testing when we implemented them. I'm again wondering if this is a deal breaker for you.
