### 2024-01-23 American-Institutes-for-Research/EdSurvey#64 @wjakethompson
Hi--sorry for the delayed response. No, this is not a deal breaker for me. I tend to think of goodpractice output as best practice in *most* scenarios. But if there's a reason for the complexity to be higher (which it sounds like there is in that case), then I don't think goodpractice should be binding.
