### 2024-04-11 #5835 @jrosen48
## Conflict of interest

General checks

- [ ]  Repository: Is the source code for this software available at the https://github.com/American-Institutes-for-Research/EdSurvey/?
- [ ]  License: Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
- [ ]  Contribution and authorship: Has the submitting author (@pdbailey0) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
- [ ] Substantial scholarly effort: Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
- [ ]  Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.
- [ ]  Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.
- [ ]  Human and animal research: If the paper contains original data research on humans subjects or animals, does it comply with [JOSS's human participants research policy and/or animal research policy](https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies)? If the paper contains no such data, please check this item.

- [ ] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

 - [ ] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
- [ ]  A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?
- [ ]  State of the field: Do the authors describe how this software compares to other commonly-used packages?
 Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
- [ ]  References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?
