### 2024-04-11 #5835 @jrosen48
I've also completed my review. A few additional notes:

1. I filed an issue here: https://github.com/American-Institutes-for-Research/EdSurvey/issues
2. I am having a bit of a hard time understanding the nature and extent of the automated tests (in part because there seem to be many - a good issue)! Is it possible to separate the files with the tests from the data/output used as part of them ([here](https://github.com/American-Institutes-for-Research/EdSurvey/tree/main/tests/testthat))? Alternatively, could the testing coverage be automatically displayed via https://covr.r-lib.org?
3. While the paper is well-written and the diagram for the workflow is important and valuable, could the paper include a minimal example of the code used to access and conduct a minimal analysis using the package?
