### 2024-05-06 #5835 @yuqiliao
Hi @jrosen48 , thanks for your review and feedback!

We're actively addressing your first two points. Regarding note 3 about including a minimal code example, we've followed the submission guidelines closely. As per the [guidelines](https://joss.readthedocs.io/en/latest/submitting.html), `a “full length” paper is not permitted, and software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation.`

Therefore, in our manuscript, we refrained from including code snippets but ensured comprehensive examples are included in our documentation. These examples are accessible via function calls using `?` and are also available in our online user's guide [here](https://naep-research.airprojects.org/Portals/0/EdSurvey_A_Users_Guide/_book/index.html). Would resolving note 3 without any changes be acceptable to you?
