### 2024-05-12 #5835 @jrosen48
Appreciate the response and the reasoning. Isn't a minimal example different from documentation of an API, though? Just looking for one example of one API call - not API documentation.
