### 2024-05-26 #5835 @oliviaguest
Is the reason for asking for a MWE because the package is hard tounderstand without one? If so, I am happy for more explanation or if the reviewers really think it's needed to add one. Ideally, I do think indeed that maybe example usecases are grea, but I am not sure if code itself is that useful in the paper when indeed the documentation should/could have many working examples? What is the rationale proposed for a MWE in the paper itself?
