### 2024-05-26 #5835 @jrosen48
My rationale is some code in the paper as a MWE can make the workflow and nature of how the package works clearer to users (and serve as a starting point for users' analyses using different data sources available through the package, even if the MWE illustrates only one).
