### 2024-08-01 #5835 @samhforbes
Just stepping in for @oliviaguest here.
Hi @jrosen48 and @wjakethompson is the issue that a reader of the paper might struggle to jump in and test the code from the paper and repo as seen, without referring to the full book?

Could a compromise be to include a quick example of how to get started in the README or a vignette perhaps? Then referring to the book for more details. Your thoughts @pdbailey0?
