### 2024-09-20 #5835 @samhforbes
Hi @jrosen48 and @wjakethompson just wanted to check that we are all on the same page here. Will this answer your issue adequately?
