### 2024-09-24 #5835 @wjakethompson
I didn't have any concerns with paper originally, so I'm good with this as well.
