### 2024-10-28 #5835 @pdbailey0
Thanks. I pushed several updates today. I'll update the issues on the EdSurvey GH page tomorrow. For now I'll mention that I expanded [the vignette](https://github.com/American-Institutes-for-Research/EdSurvey/blob/main/inst/doc/introduction.html).
