### 2025-01-21 #5835 @pdbailey0
I have been working on them and have resolved most of them. I'll try to update them in the next few weeks.
